feat: Add Healthmonitoring with SIGUSR1&2
The main goal of creating watchdog process was to enable health monitoring - that is if all of the processes work in a way the should.
With an use of user signals WD is sending checking whether the processes are alive,
and if any of them is not working properly it sends sigint to all of them to abort operation.
When interrupted all of the processes close their semaphores and server makes sure that shared memory segments are closed.