feat: Add Watchdog (#12)
* feat: Watchdog with SIGINT
This commit contains changes within the whole repo. The main functionality is to ensure the safety of memory across all of the processes.

* fix: Add delay to prevent processes from crashing

* feat: Add Healthmonitoring with SIGUSR1&2

The main goal of creating watchdog process was to enable health monitoring - that is if all of the processes work in a way the should.
With an use of user signals WD is sending checking whether the processes are alive,
and if any of them is not working properly it sends sigint to all of them to abort operation.
When interrupted all of the processes close their semaphores and server makes sure that shared memory segments are closed.