fix the bug with drone speed increasing infinetly
When user provided input, keyboard manager was not cleaning up the buffer.
Keyboard manager was still reading the same button, until il new one appeared.
What was happening, instead of cleaning up the value of shared memory,
the file descriptor for that memory was being set 0.

The solution is simply to set the value of the shared memory.