feat: Info logs
This commit consists of minor changes to all the files to allow logging
for info messages.
Logging happens through one single file, all processes open, append and
close the file when they need to.

In the end the output for the logs can be found in build/logs/ with the
date of the session.